add: Domain Entity 파일 분리
