+++ iOS/Projects/Features/Record/Sources/Domain/Entities/KalmanFilterCensored.swift
@@ -1,0 +1,16 @@
+//
+//  KalmanFilterCensored.swift
+//  RecordFeature
+//
+//  Created by MaraMincho on 12/4/23.
+//  Copyright © 2023 kr.codesquad.boostcamp8. All rights reserved.
+//
+
+import Foundation
+
+// MARK: - KalmanFilterCensored
+
+public struct KalmanFilterCensored {
+  let longitude: Double
+  let latitude: Double
+}
