+++ iOS/Projects/Features/Record/Sources/Domain/Entities/KalmanFilterUpdateRequireElement.swift
@@ -1,0 +1,18 @@
+//
+//  KalmanFilterUpdateRequireElement.swift
+//  RecordFeature
+//
+//  Created by MaraMincho on 12/4/23.
+//  Copyright © 2023 kr.codesquad.boostcamp8. All rights reserved.
+//
+
+import Foundation
+
+// MARK: - KalmanFilterUpdateRequireElement
+
+struct KalmanFilterUpdateRequireElement {
+  let longitude: Double
+  let latitude: Double
+  let prevSpeedAtLongitude: Double
+  let prevSpeedAtLatitude: Double
+}
